string_nconcat

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -38,12 +38,12 @@ char *string_nconcat(char *s1, char *s2, unsigned int n)
 	len1 = _strlen(s1);
 	len2 = _strlen(s2);
 
-	if (n >= len2)
+	if ((int)n >= len2)
 	{
 		n = len2;
 	}
 
-	if (n < 0)
+	if ((int)n < 0)
 	{
 		return (NULL);
 	}
